requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
google-generativeai==0.8.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
Pillow==10.4.0
aiohttp==3.10.5
aiofiles==24.1.0
